Client Names & Request Numbers
Added client names + request numbers to satisfy the logging requirements, and
also fixed some issues with the ports not being available.

+++ detector/detector.py
@@ -101,7 +101,7 @@ if len(sys.argv) < 2:
 if len(sys.argv) > 2:
     raise ValueError("Too Many CLI Arguments!")
  
-freq = int(sys.argv[1])
+freq = float(sys.argv[1])
 
 while True:
